Deal with disappearing items during upload (Issue: #258)
There are several places where a disappearing directory of file might
be stat-ed for file size etc, raising an exception.

If the entry is removed between the creation and the run of the monitor
loop, addRecursive is called on a non-existing path, deal with this.

If the entry is a directory is removed after the directory entry is
created but before the children are dealt with, dirEntries raises an
exception, deal with this.

If a file disappears during the upload, the upload process has a file
handle open, so can finish the upload, but the following file size
computation and time stamp update fails. The file size computation
can be fixed by re-using the already done call to getSize before the
file upload, the time stamp update is again protected with checking
for the existence of the file.