get rid of forced root necessity during installation
Currently, we require installation to be carried out by root:
- we create /var/log/onedrive with ownership root.users
- we create the systemd unit directories with root.root

Remove the creation of /var/log/onedrive directory. The default is
not to log, and fall back to logging to home directory. It is
the obligation of the system administrator to create these
directory. Furthermore, we don't know whether the group "users"
is actually present.

Remove the owner/group specification from the directory creation.
We install the actual service files without owner/group specs
anyway.

Update USAGE.md to document this.