systemd: disable non-working remaining sandboxing options
ProtectSystem= and other sandboxing options require a user namespace in
order to work as user units (the user manager does not run as root and
thus without a user namespace it is unable to perform mounts).
Add a commented out PrivateUsers=yes and disable the remaining
non-working sandoxing options that were still enabled.