Fix RC2 pthread_mutex_destroy failed segfault
When using 'onedrive-pr2709' up to https://github.com/abraunegg/onedrive/pull/2709/commits/fb77411b1fb0c010209306668cd6d88480eee4b4 the client will segfault with the following error when running in --monitor mode and the client is exited with CTRL-C:

Aborting from core/sync/mutex.d(149) Error: pthread_mutex_destroy failed.Aborted

This has been traced back to #2582 which was a manual addition of #2519

Both #2582 and #2519 segfault on exit, with #2519 just segfaulting, but #2582 generating the above error.

As such, this PR is a removal of #2519, #2609 and #2628.

When this specific PR is running, in comparison to RC2, there is zero segfault on exit when running in --monitor mode and using CTRL-C to exit.